Implement permit functionality  (#340)
* add ERC20 permit contracts

* add ERC20PermitUpgradeable

* remove erc20upgradeable

* add back ERC20Upgradeable

* fix build

+++ specs/harness/OtokenHarnessB.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.6.0;
-import "../../contracts/Otoken.sol";
+import '../../contracts/Otoken.sol';
 
 contract OtokenHarnessB is Otoken {
-    function havocTotalSupply(uint newVal) public {
-        _totalSupply = newVal;
-    }
+  function havocTotalSupply(uint256 newVal) public {
+    _totalSupply = newVal;
+  }
 }
